Refactoring WorkspaceService.createWorkspace - Created new WorkspaceCreatorSpec and moved all existing tests out of WorkspaceServiceSpec into the new spec file.  Those tests are passing.  Next, cloneWorkspace and then removing the old code from WorkspaceService
